# docs/openapi.yaml
openapi: 3.0.0
info:
  title: IAM Service API
  description: API documentation for the Identity and Access Management (IAM) service, covering Redirect, Roles, and Users modules.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /redirect/{institution_id}:
    get:
      summary: Handle IDP Redirect
      description: Handles the redirect flow from an Identity Provider. Validates the token, creates a session, and redirects the user.
      tags:
        - Redirect
      security: [] # Public endpoint
      parameters:
        - in: path
          name: institution_id
          schema:
            type: string
          required: true
          description: ID of the institution
        - in: query
          name: token
          schema:
            type: string
          required: true
          description: IDP access token
      responses:
        '302':
          description: Redirects to the institution's configured dashboard URL with a session cookie.
        '400':
          description: Bad Request (missing params)
        '403':
          description: Forbidden (invalid token)
        '404':
          description: Not Found (institution or user)

  # Roles Module
...